requirements.txt
Flask==3.0.3
weasyprint==62.3
waitress==3.0.0
Flask-sqlalchemy==3.1.1
vietnam_number
flask-session==0.8.0